requirements.txt
tensorflow-gpu==1.14
numpy==1.18.5
scikit-learn==0.23.2
seaborn==0.11.0
pandas==1.1.3
matplotlib==3.3.2